Add CheckStyle to Gradle project
This commit adds CheckStyle to the Gradle project. The CheckStyle plugin has been added to the build.gradle.kts file and the CheckStyle configuration file has been added to the project. The configuration file has been customized to match the project's coding standards. This change ensures that the code in the project is consistent with the coding standards and improves the overall quality of the code.